requirements.txt
ttkbootstrap~=1.10.1
pillow~=11.1.0
pytablericons~=1.0.1